+++ examples/AutoBenchmark/generate_readme.py
@@ -106,7 +106,7 @@ should be pretty close to the time taken by the function if they were isolated.
 ### Arduino Nano
 
 * 16MHz ATmega328P
-* Arduino IDE 1.8.13
+* Arduino IDE 1.8.16, Arduino CLI 0.19.2
 * Arduino AVR Boards 1.8.3
 * `micros()` has a resolution of 4 microseconds
 
@@ -117,7 +117,7 @@ should be pretty close to the time taken by the function if they were isolated.
 ### SparkFun Pro Micro
 
 * 16 MHz ATmega32U4
-* Arduino IDE 1.8.13
+* Arduino IDE 1.8.16, Arduino CLI 0.19.2
 * SparkFun AVR Boards 1.1.13
 * `micros()` has a resolution of 4 microseconds
 
@@ -128,8 +128,8 @@ should be pretty close to the time taken by the function if they were isolated.
 ### SAMD21 M0 Mini
 
 * 48 MHz ARM Cortex-M0+
-* Arduino IDE 1.8.13
-* SparkFun SAMD Core 1.8.3
+* Arduino IDE 1.8.16, Arduino CLI 0.19.2
+* SparkFun SAMD Core 1.8.5
 
 ```
 {samd_results}
@@ -138,7 +138,7 @@ should be pretty close to the time taken by the function if they were isolated.
 ### STM32
 
 * STM32 "Blue Pill", STM32F103C8, 72 MHz ARM Cortex-M3
-* Arduino IDE 1.8.13
+* Arduino IDE 1.8.16, Arduino CLI 0.19.2
 * STM32duino 2.0.0
 
 ```
@@ -148,8 +148,8 @@ should be pretty close to the time taken by the function if they were isolated.
 ### ESP8266
 
 * NodeMCU 1.0 clone, 80MHz ESP8266
-* Arduino IDE 1.8.13
-* ESP8266 Boards 2.7.4
+* Arduino IDE 1.8.16, Arduino CLI 0.19.2
+* ESP8266 Boards 3.0.2
 
 ```
 {esp8266_results}
@@ -158,7 +158,7 @@ should be pretty close to the time taken by the function if they were isolated.
 ### ESP32
 
 * ESP32-01 Dev Board, 240 MHz Tensilica LX6
-* Arduino IDE 1.8.13
+* Arduino IDE 1.8.16, Arduino CLI 0.19.2
 * ESP32 Boards 1.0.6
 
 ```
@@ -168,8 +168,8 @@ should be pretty close to the time taken by the function if they were isolated.
 ### Teensy 3.2
 
 * 96 MHz ARM Cortex-M4
-* Arduino IDE 1.8.13
-* Teensyduino 1.53
+* Arduino IDE 1.8.16, Arduino CLI 0.19.2
+* Teensyduino 1.55
 * Compiler options: "Faster"
 
 ```