+++ examples/MemoryBenchmark/generate_readme.py
@@ -29,7 +29,7 @@ memory and static RAM sizes were recorded. The `FEATURE_BASELINE` selection is
 the baseline, and its memory usage numbers are subtracted from the subsequent
 `FEATURE_*` memory usage.
 
-**Version**: AceCommon v1.4.6
+**Version**: AceCommon v1.4.7
 
 **DO NOT EDIT**: This file was auto-generated using `make README.md`.
 
@@ -80,6 +80,15 @@ ASCII table.
 * Add benchmark for `isSorted()`.
     * Add `array[100]` into Baseline to compensate for its use in `isSorted()`.
 * Add benchmark for `printUint16AsFloat3To()` and `printUint32AsFloat3To()`.
+
+**v1.4.7**
+
+* Add benchmarks for `KString::compareTo(const KString&)`.
+* Upgrade Arduino IDE from 1.8.13 to 1.8.16.
+* Upgrade Arduino CLI from 0.14.0 to 0.19.2
+* Upgrade SparkFun SAMD Core from 1.8.3 to 1.8.5.
+* Upgrade ESP8266 Core from 2.7.4 to 3.0.2.
+* Upgrade Teensyduino from 1.53 to 1.55.
 
 ## Arduino Nano